"Added wins column to ORDER BY clause in createGroupTableResultsView function"
The code diff shows the addition of the `wins DESC` line to the `ORDER BY` clause in the `createGroupTableResultsView` function. This change will result in the group table results being ordered by the number of wins in descending order, followed by the number of sets, total games, and games won (in descending order as well).