"Added userRole prop to TournamentsTable component and updated related functions in data.ts"
This change adds a new prop called `userRole` to the `TournamentsTable` component and updates related functions in `data.ts` to accept this new prop. However, the actual implementation of the `userRole` prop is not visible in this diff, as it only shows the changes made to the code. Therefore, it is not possible to provide a more specific commit message without additional context.